data_ingestion/requirements.txt
kafka-python==2.3.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
requests==2.32.5
websocket-client==1.9.0
